Wed, May 24, 2023, 7:01 PM +03:00

+++ interpreter/interpreter.py
@@ -1,8 +1,8 @@
 def main():
     expression = input("Expression: ")
     x, y, z = expression.split(" ")
-    float(x)
-    float(z)
+    int(x)
+    int(z)
     if y == "+":
         result = float(x + z)
         #round(result,1)
